readme docker additions (#6)
